refactor: (#56) shared 레이어 index.ts 배럴 파일 생성 및 import 경로 일괄 수정
