+++ src/shared/ui/ErrorMessage/index.ts
@@ -1,0 +1,1 @@
+export { ErrorMessage } from './ErrorMessage';
