+++ src/shared/ui/FormField/FormField.tsx
@@ -1,5 +1,6 @@
 import type { ReactNode } from 'react';
-import { cn } from '@/shared/lib/cn';
+
+import { cn } from '@/shared/lib';
 
 interface FormFieldProps {
   isFocused: boolean;
