+++ src/shared/ui/FormField/index.ts
@@ -1,0 +1,1 @@
+export { FormField } from './FormField';
