+++ src/shared/ui/Leaf/index.ts
@@ -1,0 +1,1 @@
+export { Leaf } from './Leaf';
