+++ src/shared/ui/RootLayout/index.ts
@@ -1,0 +1,1 @@
+export { RootLayout } from './RootLayout';
