+++ src/shared/ui/Spinner/Spinner.utils.ts
