+++ src/shared/ui/ValidationIcon/ValidationIcon.stories.tsx
