+++ src/shared/ui/ValidationIcon/ValidationIcon.tsx
