--- src/shared/ui/index.ts
@@ -1,8 +1,0 @@
-export * from './Button';
-export * from './ErrorMessage';
-export * from './FormField';
-export * from './Input';
-export * from './RootLayout';
-export * from './Spinner';
-export * from './ValidationIcon';
-export * from './Leaf';
